boombox: integrate ID12LA and make tags play songs
Instead of playing tracks in the background using interrupts, we
feed the vs1053 buffer in the loop(). The reason for this is that
feeding the buffer from interrupt context is blocking all other
interrupts for quite a long time, which will for example cause
the SoftwareSerial library to read bad data.