boombox: add self-powerdown functionality
The arduino has a self-powerdown circuit attached to it. This will
power the arduino via the 5V pin when an external button gets
pushed, and will continue to power it as long as the arduino holds
high pin 10. If pin 10 gets set to low, power will no longer be
supplied to the 5V pin and the arduino will shut down.

As soon as the sketch is initialized pin 10 will be set to high.
The boombox will shut down if it was sitting idle for more than
30 seconds after power-up or if a track stopped playing by
reaching its end.